feat: add capability to perform detection through the UI
This change includes moving all the core functionality to detection_window.py
in order to more tightly integrate it with the QDialog through signals.

It also comments out the "file" and "files" file browser widgets
in the main UI to clean it up a bit.

The model has been moved from Core to Globals.

The detection module now takes a "notify_progress" callback to notify the QThread
worker of progress when a single batch has been finished.