ci: add caching through action/cache@v2 (#51)
This change removes the system markers when installing torch and torchvision because poetry doesn't understand pytorch's use of '+' for tagging. None of the about 100 issues seem to have a perfect solution for this so I ended up opting for "poe" which is a tool that can be used to run commands through poetry.

This change also adds caching through action/cache@v2.