feat: improve video throughput in frame grabber (closes #57)
This change includes moving video preparation (letterboxing, padding etc) to the frame grabber.
The frame grabber now spawns 5 threads:
* 1 for constantly grabbing frames directly from the video (no longer using "LoadImages")
and loading them into batches.
* 4 for processing and preparing the loaded batches when notified by the frame grabber thread.

This change saw a ~100% performance increase on my setup.
A 2 minute video now takes ~10 seconds instead of ~20 seconds with the old setup.
We are now able to load larger batches into the GPU-
because it is no longer waiting for the frame grabber to get images-
and we no longer do image preparation in a single thread before detection-
which overall improves throughput and GPU utilization.