feat: add script to generate a final dataset
This script takes output cvat annotation dumps in the cvat video 1.1 format and turns them all
intoindividual "datasets" in folders with the name of the source videos they came from. Then it
createsthe training and validation splits by splitting it into train.txt and val.txt that points to
theimage files. It then splits a yolo dump (non video format) from the previous nina group and
mergesboth into a new folder with a newmerged train.txt and val.txt and also creates a
dataset.ymlcontaining the configuration data for the dataset.